backend-core/requirements.txt
Django
djangorestframework
django-cors-headers
celery
redis
requests